fix return type of method getWorlDim

+++ src/main/java/it/unibo/game/app/model/ModelImpl.java
@@ -109,7 +109,7 @@ public class ModelImpl implements Model{
     }
 
     @Override
-    public Pair<Integer, Integer> getWorldDim() {
+    public Pair<Double, Double> getWorldDim() {
         return SizeCalculation.getWorldSize();
     }
